JAV-487 extracted rest template transport
