+++ transports/transport-resttemplate/src/main/java/io/servicecomb/saga/transports/resttemplate/RestTemplateConfig.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2017 Huawei Technologies Co., Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.servicecomb.saga.transports.resttemplate;
+
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+import io.servicecomb.saga.transports.RestTransport;
+
+@Configuration
+public class RestTemplateConfig {
+
+  @LoadBalanced
+  @Bean
+  RestTemplate restTemplate() {
+    return new RestTemplate();
+  }
+
+  @Bean
+  RestTransport restTransport(RestTemplate restTemplate) {
+    return new RestTemplateTransport(restTemplate, "http://");
+  }
+}
